refactor server side data code
