+++ lib/bootevent.js
@@ -33,7 +33,8 @@ function boot (env) {
     ctx.food = require('./food')(env, ctx);
     ctx.pebble = require('./pebble')(env, ctx);
     ctx.bus = require('./bus')(env.settings, ctx);
-    ctx.data = require('./data')(env, ctx);
+    ctx.ddata = require('./data/ddata')(env, ctx);
+    ctx.dataloader = require('./data/dataloader')(env, ctx);
     ctx.notifications = require('./notifications')(env, ctx);
 
     next( );
@@ -52,7 +53,7 @@ function boot (env) {
 
   function setupListeners (ctx, next) {
     var updateData = _.debounce(function debouncedUpdateData ( ) {
-      ctx.data.update(function dataUpdated () {
+      ctx.dataloader.update(ctx.ddata, function dataUpdated () {
         ctx.bus.emit('data-loaded');
       });
     }, UPDATE_THROTTLE);
