+++ lib/notifications.js
@@ -59,12 +59,12 @@ function init (env, ctx) {
     //TODO: Announcement hack a1/a2
     var level = notify.isAnnouncement ? 'a' + notify.level : notify.level;
     var alarm = getAlarm(level);
-    if (ctx.data.lastUpdated > alarm.lastAckTime + alarm.silenceTime) {
+    if (ctx.ddata.lastUpdated > alarm.lastAckTime + alarm.silenceTime) {
       ctx.bus.emit('notification', notify);
-      alarm.lastEmitTime = ctx.data.lastUpdated;
+      alarm.lastEmitTime = ctx.ddata.lastUpdated;
       logEmitEvent(notify);
     } else {
-      console.log(alarm.label + ' alarm is silenced for ' + Math.floor((alarm.silenceTime - (ctx.data.lastUpdated - alarm.lastAckTime)) / 60000) + ' minutes more');
+      console.log(alarm.label + ' alarm is silenced for ' + Math.floor((alarm.silenceTime - (ctx.ddata.lastUpdated - alarm.lastAckTime)) / 60000) + ' minutes more');
     }
   }
