+++ lib/sandbox.js
@@ -46,12 +46,12 @@ function init ( ) {
 
     sbx.time = Date.now();
     sbx.settings = env.settings;
-    sbx.data = ctx.data.clone();
+    sbx.data = ctx.ddata.clone();
     sbx.notifications = safeNotifications(ctx);
 
     var profile = require('./profilefunctions')();
     //Plugins will expect the right profile based on time
-    profile.loadData(ctx.data.profiles);
+    profile.loadData(ctx.ddata.profiles);
     sbx.data.profile = profile;
     delete sbx.data.profiles;
