+++ lib/websocket.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var levels = require('./levels');
+var calcData = require('./data/calcdelta');
 
 function init (env, ctx, server) {
 
@@ -45,7 +46,7 @@ function init (env, ctx, server) {
 
   function emitData (delta) {
     if (lastData.cals) {
-      console.log('running websocket.emitData', ctx.data.lastUpdated);
+      console.log('running websocket.emitData', ctx.ddata.lastUpdated);
       io.to('DataReceivers').emit('dataUpdate', delta);
     }
   }
@@ -154,14 +155,14 @@ function init (env, ctx, server) {
   websocket.update = function update ( ) {
     console.log('running websocket.update');
     if (lastData.sgvs) {
-      var delta = ctx.data.calculateDelta(lastData);
+      var delta = calcData(lastData, ctx.ddata);
       if (delta.delta) {
         console.log('lastData full size', JSON.stringify(lastData).length,'bytes');
         if (delta.sgvs) { console.log('patientData update size', JSON.stringify(delta).length,'bytes'); }
         emitData(delta);
       } else { console.log('delta calculation indicates no new data is present'); }
     }
-    lastData = ctx.data.clone();
+    lastData = ctx.ddata.clone();
   };
 
   websocket.emitNotification = function emitNotification (notify) {
