+++ tests/ar2.test.js
@@ -13,7 +13,7 @@ describe('ar2', function ( ) {
 
   var env = require('../env')();
   var ctx = {};
-  ctx.data = require('../lib/data')(env, ctx);
+  ctx.ddata = require('../lib/data/ddata')();
   ctx.notifications = require('../lib/notifications')(env, ctx);
 
   var now = Date.now();
@@ -40,14 +40,14 @@ describe('ar2', function ( ) {
   }
 
   it('should plot a cone', function () {
-    ctx.data.sgvs = [{mgdl: 100, mills: before}, {mgdl: 105, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 100, mills: before}, {mgdl: 105, mills: now}];
     var sbx = prepareSandbox();
     var cone = ar2.forecastCone(sbx);
     cone.length.should.equal(26);
   });
 
   it('should plot a line if coneFactor is 0', function () {
-    ctx.data.sgvs = [{mgdl: 100, mills: before}, {mgdl: 105, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 100, mills: before}, {mgdl: 105, mills: now}];
 
     var env0 = require('../env')();
     env0.extendedSettings = { ar2: { coneFactor: 0 } };
@@ -60,7 +60,7 @@ describe('ar2', function ( ) {
 
   it('Not trigger an alarm when in range', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mgdl: 100, mills: before}, {mgdl: 105, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 100, mills: before}, {mgdl: 105, mills: now}];
 
     var sbx = prepareSandbox();
     ar2.checkNotifications(sbx);
@@ -71,7 +71,7 @@ describe('ar2', function ( ) {
 
   it('should trigger a warning when going above target', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mgdl: 150, mills: before}, {mgdl: 170, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 150, mills: before}, {mgdl: 170, mills: now}];
 
     var sbx = prepareSandbox();
     sbx.offerProperty('iob', function setFakeIOB() {
@@ -91,7 +91,7 @@ describe('ar2', function ( ) {
 
   it('should trigger a urgent alarm when going high fast', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mgdl: 140, mills: before}, {mgdl: 200, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 140, mills: before}, {mgdl: 200, mills: now}];
 
     var sbx = prepareSandbox();
     ar2.checkNotifications(sbx);
@@ -104,7 +104,7 @@ describe('ar2', function ( ) {
 
   it('should trigger a warning when below target', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mgdl: 90, mills: before}, {mgdl: 80, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 90, mills: before}, {mgdl: 80, mills: now}];
 
     var sbx = prepareSandbox();
     ar2.checkNotifications(sbx);
@@ -117,7 +117,7 @@ describe('ar2', function ( ) {
 
   it('should trigger a warning when almost below target', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mgdl: 90, mills: before}, {mgdl: 83, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 90, mills: before}, {mgdl: 83, mills: now}];
 
     var sbx = prepareSandbox();
     ar2.checkNotifications(sbx);
@@ -130,7 +130,7 @@ describe('ar2', function ( ) {
 
   it('should trigger a urgent alarm when falling fast', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mgdl: 120, mills: before}, {mgdl: 85, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 120, mills: before}, {mgdl: 85, mills: now}];
 
     var sbx = prepareSandbox();
     ar2.checkNotifications(sbx);
@@ -145,7 +145,7 @@ describe('ar2', function ( ) {
     ctx.notifications.initRequests();
 
     //same as previous test but prev is 10 mins ago, so delta isn't enough to trigger an urgent alarm
-    ctx.data.sgvs = [{mgdl: 120, mills: before - SIX_MINS}, {mgdl: 85, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 120, mills: before - SIX_MINS}, {mgdl: 85, mills: now}];
 
     var sbx = prepareSandbox();
     ar2.checkNotifications(sbx);
@@ -158,8 +158,8 @@ describe('ar2', function ( ) {
 
   it('should include current raw bg and raw bg forecast when predicting w/raw', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{unfiltered: 113680, filtered: 111232, mgdl: 100, mills: before, noise: 1}, {unfiltered: 183680, filtered: 111232, mgdl: 100, mills: now, noise: 1}];
-    ctx.data.cals = [{scale: 1, intercept: 25717.82377004309, slope: 766.895601715918, mills: now}];
+    ctx.ddata.sgvs = [{unfiltered: 113680, filtered: 111232, mgdl: 100, mills: before, noise: 1}, {unfiltered: 183680, filtered: 111232, mgdl: 100, mills: now, noise: 1}];
+    ctx.ddata.cals = [{scale: 1, intercept: 25717.82377004309, slope: 766.895601715918, mills: now}];
 
     var envRaw = require('../env')();
     envRaw.extendedSettings = {'ar2': {useRaw: true}};
@@ -190,8 +190,8 @@ describe('ar2', function ( ) {
 
   it('should not trigger an alarm when raw is missing or 0', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{unfiltered: 0, filtered: 0, mgdl: 100, mills: before, noise: 1}, {unfiltered: 0, filtered: 0, mgdl: 100, mills: now, noise: 1}];
-    ctx.data.cals = [{scale: 1, intercept: 25717.82377004309, slope: 766.895601715918, mills: now}];
+    ctx.ddata.sgvs = [{unfiltered: 0, filtered: 0, mgdl: 100, mills: before, noise: 1}, {unfiltered: 0, filtered: 0, mgdl: 100, mills: now, noise: 1}];
+    ctx.ddata.cals = [{scale: 1, intercept: 25717.82377004309, slope: 766.895601715918, mills: now}];
 
     var sbx = rawSandbox();
     ar2.checkNotifications(sbx.withExtendedSettings(ar2));
@@ -203,8 +203,8 @@ describe('ar2', function ( ) {
 
   it('should trigger a warning (no urgent for raw) when raw is falling really fast, but sgv is steady', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{unfiltered: 113680, filtered: 111232, mgdl: 100, mills: before, noise: 1}, {unfiltered: 43680, filtered: 111232, mgdl: 100, mills: now, noise: 1}];
-    ctx.data.cals = [{scale: 1, intercept: 25717.82377004309, slope: 766.895601715918, mills: now}];
+    ctx.ddata.sgvs = [{unfiltered: 113680, filtered: 111232, mgdl: 100, mills: before, noise: 1}, {unfiltered: 43680, filtered: 111232, mgdl: 100, mills: now, noise: 1}];
+    ctx.ddata.cals = [{scale: 1, intercept: 25717.82377004309, slope: 766.895601715918, mills: now}];
 
     var sbx = rawSandbox();
     sbx.offerProperty('rawbg', function setFakeIOB() {
@@ -221,8 +221,8 @@ describe('ar2', function ( ) {
 
   it('should trigger a warning (no urgent for raw) when raw is rising really fast, but sgv is steady', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{unfiltered: 113680, filtered: 111232, mgdl: 100, mills: before, noise: 1}, {unfiltered: 183680, filtered: 111232, mgdl: 100, mills: now, noise: 1}];
-    ctx.data.cals = [{scale: 1, intercept: 25717.82377004309, slope: 766.895601715918, mills: now}];
+    ctx.ddata.sgvs = [{unfiltered: 113680, filtered: 111232, mgdl: 100, mills: before, noise: 1}, {unfiltered: 183680, filtered: 111232, mgdl: 100, mills: now, noise: 1}];
+    ctx.ddata.cals = [{scale: 1, intercept: 25717.82377004309, slope: 766.895601715918, mills: now}];
 
     var sbx = rawSandbox();
     sbx.offerProperty('rawbg', function setFakeIOB() {
