+++ tests/errorcodes.test.js
@@ -9,13 +9,13 @@ describe('errorcodes', function ( ) {
   var now = Date.now();
   var env = require('../env')();
   var ctx = {};
-  ctx.data = require('../lib/data')(env, ctx);
+  ctx.ddata = require('../lib/data/ddata')();
   ctx.notifications = require('../lib/notifications')(env, ctx);
 
 
   it('Not trigger an alarm when in range', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mgdl: 100, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 100, mills: now}];
 
     var sbx = require('../lib/sandbox')().serverInit(env, ctx);
     errorcodes.checkNotifications(sbx);
@@ -26,7 +26,7 @@ describe('errorcodes', function ( ) {
 
   it('should trigger a urgent alarm when ???', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mgdl: 10, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 10, mills: now}];
 
     var sbx = require('../lib/sandbox')().serverInit(env, ctx);
     errorcodes.checkNotifications(sbx);
@@ -37,7 +37,7 @@ describe('errorcodes', function ( ) {
 
   it('should trigger a urgent alarm when hourglass', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mgdl: 9, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 9, mills: now}];
 
     var sbx = require('../lib/sandbox')().serverInit(env, ctx);
     errorcodes.checkNotifications(sbx);
@@ -50,7 +50,7 @@ describe('errorcodes', function ( ) {
 
   it('should trigger a low notification when needing calibration', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mgdl: 5, mills: now}];
+    ctx.ddata.sgvs = [{mgdl: 5, mills: now}];
 
     var sbx = require('../lib/sandbox')().serverInit(env, ctx);
     errorcodes.checkNotifications(sbx);
@@ -66,7 +66,7 @@ describe('errorcodes', function ( ) {
 
     for (var i = 1; i < 9; i++) {
       ctx.notifications.initRequests();
-      ctx.data.sgvs = [{mgdl: i, mills: now}];
+      ctx.ddata.sgvs = [{mgdl: i, mills: now}];
 
       var sbx = require('../lib/sandbox')().serverInit(env, ctx);
       errorcodes.checkNotifications(sbx);
