+++ tests/notifications.test.js
@@ -9,7 +9,7 @@ describe('notifications', function ( ) {
 
   var ctx = {
     bus: new Stream
-    , data: {
+    , ddata: {
       lastUpdated: Date.now()
     }
   };
