+++ tests/sandbox.test.js
@@ -33,7 +33,7 @@ describe('sandbox', function ( ) {
   function createServerSandbox() {
     var env = require('../env')();
     var ctx = {};
-    ctx.data = require('../lib/data')(env, ctx);
+    ctx.ddata = require('../lib/data/ddata')();
     ctx.notifications = require('../lib/notifications')(env, ctx);
 
     return sandbox.serverInit(env, ctx);
