+++ tests/timeago.test.js
@@ -9,7 +9,7 @@ describe('timeago', function ( ) {
   var env = require('../env')();
 
   var ctx = {};
-  ctx.data = require('../lib/data')(env, ctx);
+  ctx.ddata = require('../lib/data/ddata')();
   ctx.notifications = require('../lib/notifications')(env, ctx);
 
   function freshSBX() {
@@ -21,7 +21,7 @@ describe('timeago', function ( ) {
 
   it('Not trigger an alarm when data is current', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mills: Date.now(), mgdl: 100, type: 'sgv'}];
+    ctx.ddata.sgvs = [{mills: Date.now(), mgdl: 100, type: 'sgv'}];
 
     var sbx = freshSBX();
     timeago.checkNotifications(sbx);
@@ -32,7 +32,7 @@ describe('timeago', function ( ) {
 
   it('Not trigger an alarm with future data', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mills: Date.now() + times.mins(15).msecs, mgdl: 100, type: 'sgv'}];
+    ctx.ddata.sgvs = [{mills: Date.now() + times.mins(15).msecs, mgdl: 100, type: 'sgv'}];
 
     var sbx = freshSBX();
     timeago.checkNotifications(sbx);
@@ -43,7 +43,7 @@ describe('timeago', function ( ) {
 
   it('should trigger a warning when data older than 15m', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mills: Date.now() - times.mins(16).msecs, mgdl: 100, type: 'sgv'}];
+    ctx.ddata.sgvs = [{mills: Date.now() - times.mins(16).msecs, mgdl: 100, type: 'sgv'}];
 
     var sbx = freshSBX();
     timeago.checkNotifications(sbx);
@@ -55,7 +55,7 @@ describe('timeago', function ( ) {
 
   it('should trigger an urgent alarm when data older than 30m', function (done) {
     ctx.notifications.initRequests();
-    ctx.data.sgvs = [{mills: Date.now() - times.mins(31).msecs, mgdl: 100, type: 'sgv'}];
+    ctx.ddata.sgvs = [{mills: Date.now() - times.mins(31).msecs, mgdl: 100, type: 'sgv'}];
 
     var sbx = freshSBX();
     timeago.checkNotifications(sbx);
